requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
typing_extensions==4.12.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-environ==0.11.2
mysqlclient==2.2.6
redis==5.2.0
